Forest Area Problem
This is one of the most frequently asked interview question on Graphs using DFS and BFS Algorithm. The Problem statement and one sample test case is also given inside the file. A very good question to clear out concepts on BFS and DFS as well as N Dimmensional List or Array Concepts.