Find the First Repeated Character in String
Problem Statement - Find the first repeated character in a String. We need to find the character that occurs more than once and whose index of the second occurrence is the smallest.
Input: Hello Geeks
Output: l
The solution is to run two nested loops then start traversing from the left side. For every character, check if it repeats or not, If the character repeats, increment count of repeating characters, when the count becomes n, return that repeating character.